Fix for bitfield defaults
Previously, having defaults for bitfields did not yield the correct result. This was because bitfields can be sub-byte sized and previous implementation of defaults did not take this into consideration.

I'm not 100% certain that I've done everything correct so please review the code in dl_txt_pack before admitting the pull request.