Fix for bad memcpy
We were copying the convert-instance to the out-buffer and doing an inplace memory load there. This wrote outside the buffer because the convert-instance size was larger than the output size/buffer. We now correctly convert the instance inplace in the convert-buffer and then memcpy out the converted data into the out-buffer.