Fix for bad offset for unions to union types.
Previously, the offset calculated for the type that is set for unions was aligned by 4 bytes. This alignment must be 8 bytes for unions with structs that require 8 byte alignment. This is fixed both in the type library reader and the text packer. Also added a few unit tests to cover this and some more.