Fix for bad offset for unions to union types. (#96)
* Fix for bad offset for unions to union types.

Previously, the offset calculated for the type that is set for unions was aligned by 4 bytes. This alignment must be 8 bytes for unions with structs that require 8 byte alignment. This is fixed both in the type library reader and the text packer. Also added a few unit tests to cover this and some more.

* Fix for instance patching for previous union-alignment offset fix.

There was follow-up issues from my latest commit. This changelist should fix all of the pointer-patching functionality. I'm not sure if there's anything I missed, but if so please let me know.

* Fix for compiler warning about type-punning pointers.