added the flag 'verify' to types and members.
If that is set to 'false' no static_asserts will be generated for the type respective the members if the type is 'extern'.
Usage is for example if a member is not called the same thing as in the dl-type or as in my case where a vec3 in the
external type is built as a __m128 but I want to set it as x,y,z in dl.

Fixes issue #99 and issue #71