Added support for default-values to be used in inline-arrays if all values isn't specified or 0 for pods.
If you have an inline array of length 4 and you only write [1,2] the remaining slots will be set to:
intXX/uintXX -> 0
fp32/fp64 -> 0.0
string -> nullptr
pointer -> nullptr
array -> [] ( i.e. nullptr + size 0 )

in case of a struct it will use a completely default-initialized struct of that type, i.e. to use it
like this it requires all members of the struct to have a "default" set.

Note: there might be bugs in this functionallity since it's so new and untested... but what could possibly go wrong?

Issue #105