Fixed issue #110:
datalibrary allocates crazy amounts of memory when loading many type libraries.

when loading a binary lib we just resize the arrays to the sizes that we know we are going to need instead of speculatively
growing the arrays by doubling the size. This should both solve the issue and make the memory-behaviour slightly better in
most cases.