Correct packing of default array-of-structs
Sometimes with deep enough hierarchies the array-of-structs end up in poor offsets. This fixes that. I've tried to create a testcase for this but was unable to do so without copying basically my entire typelib and an example of it. Since that would take too much time to obfuscate, I can only add the fix for it.