Support for union enums to be indexed
Support for unions to either hash names or use member-index for union type enum.

This is default as it used to be, so fully back-compatible. You can now add ""hash_union_type" : false, " to your union and it will be 0/1/2 etc mapped for the members rather than the hashed name.