No-default-packing and pack-flags
This changelists adds a new argument to the dl_txt_pack-function, which is a field of flags called pack-flags. There are currently two flags implemented, one for not applying defaults when packing the txt - and another that assumes that there is an instance already packed where we are packing to. The latter causes the write-position for future dynamic data to be after the already packed instance, allowing for overwriting dynamic data. This does however use more memory than needed, and you can optimize the struct by unpacking/packing it.