Created failing unittest about pointers to structs
If I understand the current design then all pointers need to reference payloads in the "__subdata" section, so pointing to a member struct or a struct inside an array means the data gets duplicated.

I have no fix for this because I'm not sure which way to best tag the structs in DL's json.