Fast pointer patching plus txt_unpack for "loaded" data (#158)
* Fast pointer patching and txt_unpack for "loaded" data

For my benchmark test (8,5MB of binary data with a lot of pointers in unions), it used to take 1.92 seconds to patch the pointers on my machine. After this change it takes 0.00056 seconds to do the same thing.
I'm changing the base address which pointer offsets are calculated from to be the header address instead of the instance root. This means that DL_INSTANCE_VERSION needed to be bumped to 2. There is backwards compatibility added so old data can be read and resaved.
I'm also adding more CMake rules, to simplify building and testing on other platforms than Windows.

* Fixed GCC warnings

* Fixed bad code merge

* Changed the type of a loaded instance in the API

Co-authored-by: Christian Nilsendahl <christian.nilsendahl@elementalgames.se>